LOGIN Implemented

+++ src/main/java/com/lti/repo/RepositoryInterface.java
@@ -10,7 +10,7 @@ public interface RepositoryInterface {
 	public long registerUser(UserTable user);
 	public List<UserTable> fetchAllUsers();
 	boolean isUserPresent(String userEmail);
-	public UserTable findById(int userId);
+	public UserTable findById(long userId);
 	public UserTable findUserByUserID(long userId);
 	public long findUserIdByEmailAndPassword(String userEmail, String userPass);
 	public void removeUserByUserID(long userId);
